// generated/src/models/AsyncAcceptedResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SegmentationAPI
 * API for asynchronous image and video segmentation and dashboard management.
 *
 * The version of the OpenAPI document: 2026-08-03
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * 
 * @export
 * @interface AsyncAcceptedResponse
 */
export interface AsyncAcceptedResponse {
    /**
     * 
     * @type {string}
     * @memberof AsyncAcceptedResponse
     */
    jobId: string;
    /**
     * 
     * @type {AsyncAcceptedResponseTypeEnum}
     * @memberof AsyncAcceptedResponse
     */
    type: AsyncAcceptedResponseTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof AsyncAcceptedResponse
     */
    totalItems: number;
}


/**
 * @export
 */
export const AsyncAcceptedResponseTypeEnum = {
    Image: 'image',
    Video: 'video',
    UnknownDefaultOpenApi: '11184809'
} as const;
export type AsyncAcceptedResponseTypeEnum = typeof AsyncAcceptedResponseTypeEnum[keyof typeof AsyncAcceptedResponseTypeEnum];


/**
 * Check if a given object implements the AsyncAcceptedResponse interface.
 */
export function instanceOfAsyncAcceptedResponse(value: object): value is AsyncAcceptedResponse {
    if (!('jobId' in value) || value['jobId'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('totalItems' in value) || value['totalItems'] === undefined) return false;
    return true;
}

export function AsyncAcceptedResponseFromJSON(json: any): AsyncAcceptedResponse {
    return AsyncAcceptedResponseFromJSONTyped(json, false);
}

export function AsyncAcceptedResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AsyncAcceptedResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'jobId': json['jobId'],
        'type': json['type'],
        'totalItems': json['totalItems'],
    };
}

export function AsyncAcceptedResponseToJSON(json: any): AsyncAcceptedResponse {
    return AsyncAcceptedResponseToJSONTyped(json, false);
}

export function AsyncAcceptedResponseToJSONTyped(value?: AsyncAcceptedResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'jobId': value['jobId'],
        'type': value['type'],
        'totalItems': value['totalItems'],
    };
}