// generated/src/models/JobCreateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SegmentationAPI
 * API for asynchronous image and video segmentation and dashboard management.
 *
 * The version of the OpenAPI document: 2026-08-03
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { ImageJobCreateRequest } from './ImageJobCreateRequest.js';
import {
    instanceOfImageJobCreateRequest,
    ImageJobCreateRequestFromJSON,
    ImageJobCreateRequestFromJSONTyped,
    ImageJobCreateRequestToJSON,
} from './ImageJobCreateRequest.js';
import type { VideoJobCreateRequest } from './VideoJobCreateRequest.js';
import {
    instanceOfVideoJobCreateRequest,
    VideoJobCreateRequestFromJSON,
    VideoJobCreateRequestFromJSONTyped,
    VideoJobCreateRequestToJSON,
} from './VideoJobCreateRequest.js';

/**
 * @type JobCreateRequest
 * 
 * @export
 */
export type JobCreateRequest = { type: 'image' } & ImageJobCreateRequest | { type: 'video' } & VideoJobCreateRequest;

export function JobCreateRequestFromJSON(json: any): JobCreateRequest {
    return JobCreateRequestFromJSONTyped(json, false);
}

export function JobCreateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobCreateRequest {
    if (json == null) {
        return json;
    }
    switch (json['type']) {
        case 'image':
            return Object.assign({}, ImageJobCreateRequestFromJSONTyped(json, true), { type: 'image' } as const);
        case 'video':
            return Object.assign({}, VideoJobCreateRequestFromJSONTyped(json, true), { type: 'video' } as const);
        default:
            return json;
    }
}

export function JobCreateRequestToJSON(json: any): any {
    return JobCreateRequestToJSONTyped(json, false);
}

export function JobCreateRequestToJSONTyped(value?: JobCreateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }
    switch (value['type']) {
        case 'image':
            return Object.assign({}, ImageJobCreateRequestToJSON(value), { 'type': 'image' } as const);
        case 'video':
            return Object.assign({}, VideoJobCreateRequestToJSON(value), { 'type': 'video' } as const);
        default:
            return value;
    }
}