// generated/src/models/JobSummary.ts
export interface JobSummary {
    /**
     * 
     * @type {string}
     * @memberof JobSummary
     */
    jobId: string;
    /**
     * 
     * @type {JobSummaryTypeEnum}
     * @memberof JobSummary
     */
    type: JobSummaryTypeEnum;
    /**
     * 
     * @type {ProcessingMode}
     * @memberof JobSummary
     */
    processingMode: ProcessingMode;
    /**
     * 
     * @type {number}
     * @memberof JobSummary
     */
    totalItems: number;
    /**
     * 
     * @type {JobStatus}
     * @memberof JobSummary
     */
    status: JobStatus;
    /**
     * 
     * @type {Date}
     * @memberof JobSummary
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof JobSummary
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof JobSummary
     */
    error?: string | null;
}


/**
 * @export
 */
export const JobSummaryTypeEnum = {
    Image: 'image',
    Video: 'video',
    UnknownDefaultOpenApi: '11184809'
} as const;
export type JobSummaryTypeEnum = typeof JobSummaryTypeEnum[keyof typeof JobSummaryTypeEnum];