// generated/src/models/OutputManifest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SegmentationAPI
 * API for asynchronous image and video segmentation and dashboard management.
 *
 * The version of the OpenAPI document: 2026-08-03
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { ImageOutputManifest } from './ImageOutputManifest.js';
import {
    instanceOfImageOutputManifest,
    ImageOutputManifestFromJSON,
    ImageOutputManifestFromJSONTyped,
    ImageOutputManifestToJSON,
} from './ImageOutputManifest.js';
import type { VideoOutputManifest } from './VideoOutputManifest.js';
import {
    instanceOfVideoOutputManifest,
    VideoOutputManifestFromJSON,
    VideoOutputManifestFromJSONTyped,
    VideoOutputManifestToJSON,
} from './VideoOutputManifest.js';

/**
 * @type OutputManifest
 * 
 * @export
 */
export type OutputManifest = { type: 'image' } & ImageOutputManifest | { type: 'video' } & VideoOutputManifest;

export function OutputManifestFromJSON(json: any): OutputManifest {
    return OutputManifestFromJSONTyped(json, false);
}

export function OutputManifestFromJSONTyped(json: any, ignoreDiscriminator: boolean): OutputManifest {
    if (json == null) {
        return json;
    }
    switch (json['type']) {
        case 'image':
            return Object.assign({}, ImageOutputManifestFromJSONTyped(json, true), { type: 'image' } as const);
        case 'video':
            return Object.assign({}, VideoOutputManifestFromJSONTyped(json, true), { type: 'video' } as const);
        default:
            return json;
    }
}

export function OutputManifestToJSON(json: any): any {
    return OutputManifestToJSONTyped(json, false);
}

export function OutputManifestToJSONTyped(value?: OutputManifest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }
    switch (value['type']) {
        case 'image':
            return Object.assign({}, ImageOutputManifestToJSON(value), { 'type': 'image' } as const);
        case 'video':
            return Object.assign({}, VideoOutputManifestToJSON(value), { 'type': 'video' } as const);
        default:
            return value;
    }
}